Add files via upload
Still needs a lot of work, based on Eugene Krevenets titanic-for-beginners on Kaggle. Ran a few split train tests and pulled around 83-84% accuracy. Still need to fill in the blank Age cells and account for family and other variables. As it stands, the split train test is not in the file so you can mess with it.